Consolidate integrand_wrap_first_two_moments and integrand_trace_moments into integrand_wrap_moments
