+++ tutorials/3_uncertainty_quantification.py
@@ -26,7 +26,8 @@ num_samples = 10_000
 # Sometimes, second and higher moments of a random variable are interesting.
 
 normal = hutchinson.sampler_normal(x_like, num=num_samples)
-integrand = hutchinson.integrand_trace_moments(matvec, [1, 2])
+integrand = hutchinson.integrand_trace(matvec)
+integrand = hutchinson.integrand_wrap_moments(integrand, [1, 2])
 estimator = hutchinson.hutchinson(integrand, sample_fun=normal)
 first, second = estimator(jax.random.PRNGKey(1))
